xst: split, -l
